refactor: var names, format
