+++ main.go
@@ -1,10 +1,5 @@
 package main
 
-import (
-	"github.com/michimani/gotwi"
-	"net/http"
-)
-
 func main() {
 	client, err := newOAuth2Client()
 	if err != nil {
@@ -12,14 +7,5 @@ func main() {
 	}
 
 	//	createRules(client)
-	samplingTweets(client)
-}
-
-func newOAuth2Client() (*gotwi.Client, error) {
-	in2 := &gotwi.NewClientInput{
-		AuthenticationMethod: gotwi.AuthenMethodOAuth2BearerToken,
-		HTTPClient:           &http.Client{Timeout: 0},
-	}
-
-	return gotwi.NewClient(in2)
+	recieveStream(client)
 }
