+++ twt.go
@@ -3,51 +3,62 @@ package main
 import (
 	"context"
 	"fmt"
+	"net/http"
 
 	"github.com/michimani/gotwi"
 	"github.com/michimani/gotwi/tweet/filteredstream"
 	"github.com/michimani/gotwi/tweet/filteredstream/types"
 )
 
-func createRules(c *gotwi.Client) {
+func newOAuth2Client() (*gotwi.Client, error) {
+	in2 := &gotwi.NewClientInput{
+		AuthenticationMethod: gotwi.AuthenMethodOAuth2BearerToken,
+		HTTPClient:           &http.Client{Timeout: 0},
+	}
+
+	return gotwi.NewClient(in2)
+}
+
+func createRules(client *gotwi.Client) {
+
 	p := &types.CreateRulesInput{
 		Add: []types.AddingRule{
 			{Value: gotwi.String("参戦ID参加者募集！")},
 		},
 	}
-	s, err := filteredstream.CreateRules(context.Background(), c, p)
+
+	stream, err := filteredstream.CreateRules(context.Background(), client, p)
 	if err != nil {
 		fmt.Println(err)
 		panic(err)
 	}
-	fmt.Println(s)
+	fmt.Println(stream)
 }
 
-func samplingTweets(c *gotwi.Client) {
+func recieveStream(client *gotwi.Client) {
 
 	p := &types.SearchStreamInput{}
-	s, err := filteredstream.SearchStream(context.Background(), c, p)
+	stream, err := filteredstream.SearchStream(context.Background(), client, p)
 	if err != nil {
 		fmt.Println(err)
 		return
 	}
 
 	cnt := 0
-	for s.Receive() {
+	for stream.Receive() {
 		cnt++
-		t, err := s.Read()
+		tweet, err := stream.Read()
 		if err != nil {
 			fmt.Println(err)
 		} else {
-			if t != nil {
-
-				fmt.Println(cnt, gotwi.StringValue(t.Data.ID), gotwi.StringValue(t.Data.Text))
+			if tweet != nil {
+				fmt.Println(cnt, gotwi.StringValue(tweet.Data.ID), gotwi.StringValue(tweet.Data.Text))
 			}
 		}
 	}
 }
 
 func writeLog(cnt int, id string, txt string) {
+
 	fmt.Println(cnt, id, txt)
-	return
 }
